Move smart collection state to common
