+++ packages/web/src/store/types.ts
@@ -3,6 +3,7 @@ import { RouterState } from 'connected-react-router'
 import { CommonState } from 'common/store'
 import averageColor from 'common/store/average-color/slice'
 import { CollectionsPageState } from 'common/store/pages/collection/types'
+import { SmartCollectionState } from 'common/store/pages/smart-collection/slice'
 import RemoteConfigReducer from 'common/store/remote-config/slice'
 import StemsUploadReducer from 'common/store/stems-upload/slice'
 import { CreatePlaylistModalState } from 'common/store/ui/createPlaylistModal/types'
@@ -32,7 +33,6 @@ import SavesPageState from 'pages/saved-page/store/types'
 import { SearchPageState } from 'pages/search-page/store/types'
 import SettingsPageState from 'pages/settings-page/store/types'
 import SignOnPageState from 'pages/sign-on/store/types'
-import { SmartCollectionState } from 'pages/smart-collection/store/slice'
 import trendingPlaylistsReducer from 'pages/trending-playlists/store/slice'
 import { UploadPageState } from 'pages/upload-page/store/types'
 import VisualizerReducer from 'pages/visualizer/store/slice'
